SWPY-1079 simplify the call to log_error
error_views call log_error with the exception and some additional
exception data. There is no reason, aside for when you explicitly want
to provide custom data, that log_error itself cannot parse the
exception.

log_error now gets the error_type and error_msg from the exception but
can also receive these as keyword arguments.

The tests make sure the right information is passed to the logger after
all the setup. And also test an excisting functionality of passing along
the __cause__.

SpynlExceptions can now be initialized with a message as well as a
debug_message, which will default to be the same as message.