SPAPI-73 Get the locale setting from the user
This PR introduces a number of changes. The cookie we are settings for
language is now called _LOCALE_ as per pyramid suggestions.

We no longer explicitly set in the code a default as pyramid already
has a setting for this in the ini pyramid.default_locale_name

The validation of locales is extracted so other parts of the code can
call it and do what they want with the output.