SWPY-1140 simply try catch the import of raven/newrelic
Pip actually removed the get_installed_distributions method in the
latest version. So just import and catch the import error

sentry will actually not reraise exceptions that are raised
when sending reports. So anything that DOES go wrong there will
be a code error that we need to address. So take out the
try.. except Exception.

newrelic actually returns False if it cant add extra information
so check that instead of a try except.